add dependencies and create base class for postgresql connection
