+++ fast_datapy_etl/drivers/postgresql/constants.py
@@ -1,0 +1,39 @@
+
+from enum import Enum
+
+
+class ODBC_TYPES(Enum):
+    TEXT = "SQL"
+    CHARACTER_VARYING ="SQL_VARCHAR"
+    CHARACTER ="SQL_CHAR"
+    BIT="SQL_BIT"
+    SMALLINT = "SQL_SMALLINT"
+    INTEGER = "SQL_INTEGER"
+    BIGINT ="SQL_BIGINT"
+    OID ="SQL_INTEGER"
+    SMALL_SERIAL = "SQL_SMALLINT"
+    SERIAL = "SQL_INTEGER"
+    BIG_SERIAL = "SQL_BIGINT"
+    BYTEA = "SQL_LONGVARBINARY"
+    NUMERIC = "SQL_NUMERIC"
+    DECIMAL = "SQL_DECIMAL"
+    REAL = "SQL_REAL"
+    DOUBLE_PRECISION = "SQL_DOUBLE"
+    DATE = "SQL_TYPE_DATE"
+    TIME = "SQL_TYPE_TIME"
+    TIMESTAMP = "SQL_TYPE_TIMESTAMP"
+    INTERVAL_YEAR = "SQL_INTERVAL_YEAR"
+    INTERVAL_MONTH = "SQL_INTERVAL_MONTH"
+    INTERVAL_DAY = "SQL_INTERVAL_DAY"
+    INTERVAL_HOUR = "SQL_INTERVAL_HOUR"
+    INTERVAL_MINUTE = "SQL_INTERVAL_MINUTE"
+    INTERVAL_SECOND = "SQL_INTERVAL_SECOND"
+    INTERVAL_YEAR_TO_MONTH = "SQL_INTERVAL_YEAR_TO_MONTH"
+    INTERVAL_DAY_TO_HOUR = "SQL_INTERVAL_DAY_TO_HOUR"
+    INTERVAL_DAY_TO_MINUTE = "SQL_INTERVAL_DAY_TO_MINUT"
+    INTERVAL_DAY_TO_SECOND = "SQL_INTERVAL_DAY_TO_SECOND"
+    INTERVAL_HOUR_TO_MINUTE = "SQL_INTERVAL_DAY"
+    INTERVAL_HOUR_TO_SECOND = "SQL_INTERVAL_HOUR_TO_SECOND"
+    INTERVAL_MINUTE_TO_SECOND = "SQL_INTERVAL_MINUTE_TO_SECOND"
+
+STRING_CONN = ["Driver","Port","Database", "Server"]
