+++ pyproject.toml
@@ -5,7 +5,9 @@ description = ""
 authors = ["Agajanth <santiagoisaza0@gmail.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.6"
+python = "^3.7"
+psycopg = "^3.1.1"
+pyodbc = "^4.0.34"
 
 [tool.poetry.dev-dependencies]
 pytest = "^5.2"
